refactor callback core
