+++ gradsflow/callbacks/__init__.py
@@ -11,7 +11,6 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from .callbacks import Callback
 from .export import ModelCheckpoint
 from .gpu import EmissionTrackerCallback
 from .logger import CometCallback, CSVLogger
