+++ gradsflow/core/__init__.py
@@ -13,3 +13,4 @@
 #  limitations under the License.
 
 """Core Building blocks for Auto Tasks"""
+from .callbacks import Callback
